refactor(web-react): use Grid and Stack in `UNSTABLE_FileUpload` demos
Replace manual layout divs with Grid and Stack components for consistency with the design system.
